chore: Rename component class to make easier using it in x-components
EX-7836
