+++ packages/x-tailwindcss/demo/src/index.ts
@@ -3,6 +3,6 @@ export { default as XdsIcon } from './components/xds-icon.vue';
 export { default as XdsBaseShowcase } from './components/xds-base-showcase.vue';
 export { default as XdsShowCase } from './components/xds-showcase.vue';
 export { default as XdsHighlight } from './components/xds-highlight.vue';
-export { default as XdsProductImage } from './components/xds-product-image.vue';
+export { default as XdsPicture } from './components/xds-picture.vue';
 export * from './types/types';
 export * from './utils';
