+++ packages/x-tailwindcss/src/x-tailwind-plugin/components/picture/overlay.ts
@@ -1,7 +1,7 @@
 import { TailwindHelpers } from '../../../types';
 
 /**
- * Returns the `overlay` variant of component `product-image`.
+ * Returns the `overlay` variant of component `picture`.
  *
  * @param helpers - The {@link TailwindHelpers} to generate CSS.
  * @returns The {@link CssStyleOptions} for the variant.
